[Cookie] New Component & Contract (#181)
## Description

Closes #125 

The cookie class should not be responsible for sending the headers or
pulling name/value from `$_COOKIE`. This would be the job of the cookie
manager. Cookie manager uses `$_COOKIE` to grab cookies. If a cookie
does not exist, it uses the factory to create one. Sending the
Set-Cookie header should be done be the cookie manager as well.

The Cookie class should be about to output the header value. Should be
able to do something like:

```php
header('Set-Cookie: ' . (string) $cookie);
```

## Checklist
- [x] Updated CHANGELOG files
- [x] Updated Documentation
- [x] Unit Tests Created
- [x] php-cs-fixer